==> 03-modulo03-sprint_01/regresion-lineal/10-RL-Metricas.ipynb <==
# Pair Programming Métricas Regresión Lineal
## Ana C y Ana G

En el pair programming anterior creastéis vuestro primer modelo de machine learning usando la regresion Lineal. 

Es el momento, que con vuestros datos evaluéis si es bueno haciendo predicciones. 

Los objetivo de este pairprogramming son:

- Calculéis las métricas para vuestro modelo
- Discutid los resultados de las métricas y extraed conclusiones
- Guardad los resultados de las métricas en un csv para usarlo más adelante.

# Tratamiento de datos
# -----------------------------------------------------------------------
import numpy as np
import pandas as pd

# Gráficos
# ------------------------------------------------------------------------
import matplotlib.pyplot as plt
import seaborn as sns


#  Modelado y evaluación
# -----------------------------------------------------------------------------
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error

# Primas bajas

df=pd.read_csv('../data/08-primas_categorizadas_normalizadas.csv',index_col=0)#primas bajas
df.head()

X = df.drop(['charges_Sklearn','bmi','smoker','region','sex','northeast','northwest','southeast','southwest'], axis = 1)#predictoras
y = df["charges_Sklearn"]# respuestas

X.head(2)

x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

lr = LinearRegression(n_jobs=-1)

lr.fit(x_train, y_train)

y_predict_train = lr.predict(x_train) # hacemos las predicciones
y_predict_test = lr.predict(x_test)

train_df = pd.DataFrame({'Real': y_train, 'Predicted': y_predict_train, 'Set': ['Train']*len(y_train)})
test_df  = pd.DataFrame({'Real': y_test,  'Predicted': y_predict_test,  'Set': ['Test']*len(y_test)})
resultados = pd.concat([train_df,test_df], axis = 0)
resultados.head()

resultados['residuos'] = resultados['Real'] - resultados['Predicted']
resultados.head()

resultados_metricas = {'MAE': [mean_absolute_error(y_test, y_predict_test), mean_absolute_error(y_train, y_predict_train)],
                'MSE': [mean_squared_error(y_test, y_predict_test), mean_squared_error(y_train, y_predict_train)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_predict_test)), np.sqrt(mean_squared_error(y_train, y_predict_train))],
                'R2':  [r2_score(y_test, y_predict_test), r2_score(y_train, y_predict_train)],
                 "set": ["test", "train"], 
                 "modelo": ["Linear Regresion", "LinearRegression"]}

df_resultados = pd.DataFrame(resultados_metricas)

df_resultados

---

Resultado:

- Mas del 70% de la variación de nuestra **variable respuesta ** es explicado por nuestras variables predictoras,
-  y los valor de MAE y MSE estan muy cerca del cero, segun esto seria un buen modelo predictivo, 
- pero MSE y MAE de el grupo **train**, son menores que las del grupo **test**, 
- al contrario que ocurre con los valores de R2.
- Esto podria indicar, **overfitting**, pero en nuestro caso, las diferencias son bajas, más bien, quiere decir, que tenemos un modelo bastante bueno.

---

# Primas altas

df2=pd.read_csv('../data/09-primas-altas.csv',index_col=0)# primas altas sin normalizar
df2.head()

numericas2= df2.select_dtypes(np.number)
numericas2

X = numericas2.drop(['bmi','charges'],axis=1)#predictoras
y = df2["charges"]# respuestas

x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

lr = LinearRegression(n_jobs=-1)

lr.fit(x_train, y_train)

y_predict_train = lr.predict(x_train) 
y_predict_test = lr.predict(x_test)

train_df = pd.DataFrame({'Real': y_train, 'Predicted': y_predict_train, 'Set': ['Train']*len(y_train)})
test_df  = pd.DataFrame({'Real': y_test,  'Predicted': y_predict_test,  'Set': ['Test']*len(y_test)})
resultados = pd.concat([train_df,test_df], axis = 0)
resultados.head()

resultados['residuos'] = resultados['Real'] - resultados['Predicted']
resultados.head()

resultados_metricas2 = {'MAE': [mean_absolute_error(y_test, y_predict_test), mean_absolute_error(y_train, y_predict_train)],
                'MSE': [mean_squared_error(y_test, y_predict_test), mean_squared_error(y_train, y_predict_train)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_predict_test)), np.sqrt(mean_squared_error(y_train, y_predict_train))],
                'R2':  [r2_score(y_test, y_predict_test), r2_score(y_train, y_predict_train)],
                 "set": ["test", "train"], 
                 "modelo": ["Linear Regresion", "LinearRegression"]}

df_resultados2 = pd.DataFrame(resultados_metricas2)

df_resultados2

---

Resultados:

- En este modelo vemso que tiene un bajo valor predictivo, el valor de R no supera el 25% de predicción y los valores de MAE y MSE son muy elevados. 
- Hay que tener en cuenta que estos valores no estan normalizados.
- Tendriamos ue aumentar el numero de variables predictoras, aumentar la **n** o revisar la partición del dataset inicial.

---

# Resultados para las dos particiones (primas bajas y primas altas)

df_resultados.head()

df_resultados2.head()

---

df_resultados.to_csv("../data/10-resultados_metricas.csv")

df_resultados2.to_csv("../data/10-resultados_metricas_altas.csv")
